feat: remaned and sorted .py files
