+++ cookie_eaters/code/.dvc/.gitignore
@@ -1,0 +1,3 @@
+/config.local
+/tmp
+/cache
