+++ cookie_eaters/code/data/01_setup_data.py
@@ -1,0 +1,54 @@
+import subprocess
+import pandas as pd
+import warnings
+import datetime
+import json
+import os
+import shutil
+from pprint import pprint
+
+# shutil.rmtree("./artifacts",ignore_errors=True)
+os.makedirs("artifacts", exist_ok=True)
+
+print("Created artifacts directory")
+
+#os.makedirs("artifacts", exist_ok=True)
+os.makedirs("mlruns", exist_ok=True)
+os.makedirs("mlruns/.trash", exist_ok=True)
+
+#added this to check the flow
+print("Created mlruns directory")
+
+warnings.filterwarnings('ignore')
+pd.set_option('display.float_format',lambda x: "%.3f" % x)
+
+subprocess.run(["dvc", "pull"])
+
+print("Loading training data")
+
+data = pd.read_csv("../../data/raw/raw_data.csv")
+
+print("Total rows:", data.count())
+print(data.head(5))
+
+
+max_date = "2024-01-31"
+min_date = "2024-01-01"
+
+if not max_date:
+    max_date = pd.to_datetime(datetime.datetime.now().date()).date()
+else:
+    max_date = pd.to_datetime(max_date).date()
+
+min_date = pd.to_datetime(min_date).date()
+
+# Time limit data
+data["date_part"] = pd.to_datetime(data["date_part"]).dt.date
+data = data[(data["date_part"] >= min_date) & (data["date_part"] <= max_date)]
+
+min_date = data["date_part"].min()
+max_date = data["date_part"].max()
+date_limits = {"min_date": str(min_date), "max_date": str(max_date)}
+with open("./artifacts/date_limits.json", "w") as f:
+    json.dump(date_limits, f)
+
