+++ cookie_eaters/code/data/artifacts/date_limits.json
@@ -1,0 +1,1 @@
+{"min_date": "2024-01-01", "max_date": "2024-01-31"}
