--- cookie_eaters/code/data/load_data.py
@@ -1,20 +1,0 @@
-import subprocess
-from IPython.display import display
-import pandas as pd
-
-subprocess.run(["dvc", "pull"])
-
-print("Loading training data")
-
-data = pd.read_csv("../../data/raw/raw_data.csv")
-
-print("Total rows:", data.count())
-display(data.head(5))
-
-import pandas as pd
-import warnings
-
-warnings.filterwarnings('ignore')
-pd.set_option('display.float_format',lambda x: "%.3f" % x)
-
-
