--- cookie_eaters/code/data/min_max.py
@@ -1,23 +1,0 @@
-import pandas as pd
-import datetime
-import json
-
-max_date = "2024-01-31"
-min_date = "2024-01-01"
-
-if not max_date:
-    max_date = pd.to_datetime(datetime.datetime.now().date()).date()
-else:
-    max_date = pd.to_datetime(max_date).date()
-
-min_date = pd.to_datetime(min_date).date()
-
-# Time limit data
-data["date_part"] = pd.to_datetime(data["date_part"]).dt.date
-data = data[(data["date_part"] >= min_date) & (data["date_part"] <= max_date)]
-
-min_date = data["date_part"].min()
-max_date = data["date_part"].max()
-date_limits = {"min_date": str(min_date), "max_date": str(max_date)}
-with open("./artifacts/date_limits.json", "w") as f:
-    json.dump(date_limits, f)
