+++ cookie_eaters/code/features/04_feature_engineering.py
@@ -1,3 +1,14 @@
+
+
+
+#removing irrelevant columns from the dataframe
+data = data.drop(
+    [
+        "is_active", "marketing_consent", "first_booking", "existing_customer", "last_seen", "domain", "country", "visited_learn_more_before_booking", "visited_faq"
+    ],
+    axis=1
+)
+
 #Group source values into broader categories
 
 data['bin_source'] = data['source']
