+++ cookie_eaters/code/models/10_save_model.py
@@ -1,0 +1,12 @@
+#save the model result
+column_list_path = './artifacts/columns_list.json'
+with open(column_list_path, 'w+') as columns_file:
+    columns = {'column_names': list(X_train.columns)}
+    pprint(columns)
+    json.dump(columns, columns_file)
+
+print('Saved column list to ', column_list_path)
+
+model_results_path = "./artifacts/model_results.json"
+with open(model_results_path, 'w+') as results_file:
+    json.dump(model_results, results_file)
