+++ cookie_eaters/code/models/11_mlflow_experiment.py
@@ -1,0 +1,38 @@
+import json
+import time
+from mlflow.tracking.client import MlflowClient
+from mlflow.entities.model_registry.model_version_status import ModelVersionStatus
+from mlflow.tracking.client import MlflowClient
+
+def wait_until_ready(model_name, model_version):
+    client = MlflowClient()
+    for _ in range(10):
+        model_version_details = client.get_model_version(
+          name=model_name,
+          version=model_version,
+        )
+        status = ModelVersionStatus.from_string(model_version_details.status)
+        print(f"Model status: {ModelVersionStatus.to_string(status)}")
+        if status == ModelVersionStatus.READY:
+            break
+        time.sleep(1)
+        
+#get experiment model results
+experiment_ids = [mlflow.get_experiment_by_name(experiment_name).experiment_id]
+experiment_ids
+
+experiment_best = mlflow.search_runs(
+    experiment_ids=experiment_ids,
+    order_by=["metrics.f1_score DESC"],
+    max_results=1
+).iloc[0]
+experiment_best
+
+#metrics
+with open("./artifacts/model_results.json", "r") as f:
+    model_results = json.load(f)
+results_df = pd.DataFrame({model: val["weighted avg"] for model, val in model_results.items()}).T
+results_df
+
+best_model = results_df.sort_values("f1-score", ascending=False).iloc[0].name
+print(f"Best model: {best_model}")
