+++ cookie_eaters/code/models/13_compare_register.py
@@ -1,5 +1,9 @@
 import mlflow
 import dagger
+import time
+from mlflow.tracking.client import MlflowClient
+from mlflow.entities.model_registry.model_version_status import ModelVersionStatus
+from mlflow.tracking.client import MlflowClient
 
 # Compare prod and best trained model
 train_model_score = experiment_best["metrics.f1_score"]
