--- cookie_eaters/code/models/mlplow_setup.py
@@ -1,4 +1,0 @@
-
-import mlflow
-
-mlflow.set_experiment(experiment_name)
