--- cookie_eaters/code/models/save_best_model.py
@@ -1,9 +1,0 @@
-#save the best model
-xgboost_model = model_grid.best_estimator_
-xgboost_model_path = "./artifacts/lead_model_xgboost.json"
-xgboost_model.save_model(xgboost_model_path)
-
-model_results = {
-    xgboost_model_path: classification_report(y_train, y_pred_train, output_dict=True)
-}
-
