--- cookie_eaters/code/models/train_data.py
@@ -1,10 +1,0 @@
-import json
-
-data_columns = list(data.columns)
-with open('./artifacts/columns_drift.json','w+') as f:           
-    json.dump(data_columns,f)
-    
-data.to_csv('./artifacts/training_data.csv', index=False)
-
-data.to_csv('./artifacts/train_data_gold.csv', index=False)
-
